textbook

+++ server/admin/services/manager.py
@@ -2,45 +2,36 @@ import uuid
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from admin.schema import ModifyPasswordSchema, CreateManangeSchema, UpdateManangeSchema
-from common.database import Manager, get_async_session
+from common.database import AsyncSessionLocal, Manager
 from common.schema import ManagerSchema
 from common.settings import envs
 from utils import encrypt
 from utils.time import now
 
 
-def _valid_init_manager(manager: Manager):
-    if manager.username == envs.ADMIN_USERNAME:
-        raise ValueError("初始管理员不能修改")
-
-
-def _valid_exist_manager(manager: Manager | None):
-    if manager:
-        raise ValueError("账号已经存在")
-
-
 async def init_super_manager():
-    db = get_async_session()
     """初始化超级管理员"""
-    manager = await db.scalar(select(Manager).where(Manager.username == envs.ADMIN_USERNAME))
-    if manager is not None:
-        return
+    async with AsyncSessionLocal() as db:
+        manager = await db.scalar(select(Manager).where(Manager.username == envs.ADMIN_USERNAME))
+        if manager is not None:
+            return
 
-    manager = Manager(
-        id=uuid.uuid4(),
-        username=envs.ADMIN_USERNAME,
-        password=encrypt.hash(envs.ADMIN_PASSWORD),
-        type=0,
-        status=1,
-    )
-    db.add(manager)
-    await db.commit()
+        manager = Manager(
+            id=uuid.uuid4(),
+            username=envs.ADMIN_USERNAME,
+            password=encrypt.hash(envs.ADMIN_PASSWORD),
+            type=0,
+            status=1,
+        )
+        db.add(manager)
+        await db.commit()
 
 
 async def create_manager(db: AsyncSession, params: CreateManangeSchema):
     """创建管理员"""
     manager = await db.scalar(select(Manager).where(Manager.username == params.username))
-    _valid_exist_manager(manager)
+    if manager:
+        raise ValueError("账号已经存在")
 
     password = encrypt.generate_password()
     manager = Manager(
@@ -58,7 +49,11 @@ async def create_manager(db: AsyncSession, params: CreateManangeSchema):
 async def reset_manager_password(db: AsyncSession, id: str):
     """重置管理员密码"""
     manager = await db.scalar(select(Manager).where(Manager.id == id))
-    _valid_exist_manager(manager)
+    if not manager:
+        raise ValueError("账号不存在")
+
+    if manager.username == envs.ADMIN_USERNAME:
+        raise ValueError("不能重置超级管理员密码")
 
     password = encrypt.generate_password()
     manager.password = encrypt.hash(password)
@@ -70,7 +65,8 @@ async def reset_manager_password(db: AsyncSession, id: str):
 async def update_manager_password(db: AsyncSession, id: str, params: ModifyPasswordSchema):
     """更新管理员密码"""
     manager = await db.scalar(select(Manager).where(Manager.id == id))
-    _valid_exist_manager(manager)
+    if not manager:
+        raise ValueError("账号不存在")
 
     origin = encrypt.hash(params.origin)
     if origin != manager.password:
@@ -78,14 +74,20 @@ async def update_manager_password(db: AsyncSession, id: str, params: ModifyPassw
 
     manager.password = encrypt.hash(params.password)
     manager.update_time = now()
+
+    token = encrypt.encode({"id": manager.id, "update_time": manager.update_time})
+    manager.token = token
     await db.commit()
 
 
 async def update_manager(db: AsyncSession, id: str, params: UpdateManangeSchema):
     """更新管理员信息"""
     manager = await db.scalar(select(Manager).where(Manager.id == id))
-    _valid_exist_manager(manager)
-    _valid_init_manager(manager)
+    if not manager:
+        raise ValueError("管理员不存在")
+
+    if manager.username == envs.ADMIN_USERNAME:
+        raise ValueError("不能修改超级管理员")
 
     if params.status is not None:
         manager.status = params.status
@@ -98,8 +100,11 @@ async def update_manager(db: AsyncSession, id: str, params: UpdateManangeSchema)
 async def delete_manager(db: AsyncSession, id: str):
     """删除管理员"""
     manager = await db.scalar(select(Manager).where(Manager.id == id))
-    _valid_exist_manager(manager)
-    _valid_init_manager(manager)
+    if not manager:
+        raise ValueError("管理员不存在")
+
+    if manager.username == envs.ADMIN_USERNAME:
+        raise ValueError("不能删除超级管理员")
 
     await db.delete(manager)
     await db.commit()
